% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_eight.R
\name{add_eight}
\alias{add_eight}
\title{Add eight to numeric vector}
\usage{
add_eight(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric vector
}
\description{
Add eight to numeric vector
}
\examples{
add_eight(2)
add_eight(3)
}
