% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seven.R
\name{add_seven}
\alias{add_seven}
\title{Add seven to numeric vector}
\usage{
add_seven(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric vector
}
\description{
Add seven to numeric vector
}
\examples{
add_seven(2)
add_seven(3)
}
