% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste.R
\name{exp_smooth}
\alias{exp_smooth}
\title{Smooth a time series}
\usage{
exp_smooth(y, a)
}
\arguments{
\item{y}{Time series to be smoothed}

\item{a}{Smooth parameter}
}
\value{
A time series smoothed from \code{y}
}
\description{
A function to smooth a time series
}
