% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_krs.R
\name{mean_krs}
\alias{mean_krs}
\title{Adaptative Kernel Regression Smoothing}
\usage{
mean_krs(y, x, x0, lambda)
}
\arguments{
\item{y}{A vector (dbl)}

\item{x}{A vector (dbl)}

\item{x0}{A vector (dbl)}

\item{lambda}{Bandwidth (dbl)}
}
\value{
A fitted model based on KRS method
}
\description{
Model a kernel regression model moother (KRS) based on a normal distribution.
}
\examples{
set.seed(998)
nobs <- 200
x <- runif(nobs)
y <- sin(4*pi*x^3) + rnorm(nobs, 0, 0.2)
plot(x, y)
x0 <- seq(0, 1, length.out = 1000)
fit1 <- mean_krs(y, x, x0, 0.06)
fit2 <- mean_krs(y, x, x0, 0.02)
plot(x, y, col = "grey")
lines(x0, fit1, col = 2)
lines(x0, fit2, col = 4)

}
