% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Data of motor trend car road tests}
\format{
A data frame with 32 rows and 11 variables

\describe{
\item{mpg}{Miles/(US) gallon}
\item{cyl}{Number of cylinders}
\item{disp}{Displacement}
\item{hp}{Gross horsepower}
\item{drat}{Rear axle ratio}
\item{wt}{Weight (1000 lbs)}
\item{qsec}{1/4 mile time}
\item{vs}{Engine (0 = V-shaped, 1 = straight)}
\item{am}{	Transmission (0 = automatic, 1 = manual)}
\item{gear}{Number of forward gears}
\item{carb}{Number of carburetors}
}
}
\source{
\url{https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/mtcars}
}
\usage{
mtcars
}
\description{
The data was extracted from the 1974 Motor Trend US magazine, and comprises
fuel consumption and 10 aspects of automobile design and performance for
32 automobiles (1973–74 models).
}
\keyword{datasets}
