% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mdc}
\alias{rcpp_mdc}
\title{Maximo Divisor Comum}
\usage{
rcpp_mdc(a, b)
}
\arguments{
\item{a}{Um valor (int)}

\item{b}{Umm valor (int)}
}
\value{
O MDC entre \code{a} e \code{b}
}
\description{
Maximo Divisor Comum
}
