% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker_simulation.R
\name{ricker_simulation}
\alias{ricker_simulation}
\title{Ricker Simulation Based on Ricker Model}
\usage{
ricker_simulation(n, nburn, r, y0)
}
\arguments{
\item{n}{Is the total number of time steps to be stored (int)}

\item{nburn}{Is the initial number of simulations to be discarded before the following iterations (int)}

\item{r}{Is the growth rate (dbl)}

\item{y0}{Is the initial population size (int)}
}
\value{
A tragetory from those parameters
}
\description{
A function to model population dynamics based on Ricker model
}
\examples{
nburn <- 100
n <- 50

y0_true <- 1
sig_true <- 0.1
r_true <- 10

Ntrue <- ricker_simulation(n = n, nburn = nburn, r = r_true, y0 = y0_true)
yobs <- Ntrue * exp(rnorm(n, 0, sig_true))

plot(yobs, type = 'b')
}
