% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwishart}
\alias{rwishart}
\title{Generate Random Wishart Distribution}
\usage{
rwishart(df, S)
}
\arguments{
\item{df}{An \code{int}, which gives the degrees of freedom of the Wishart (df > 0).}

\item{S}{An \code{matrix} with dimensions m x m that provides Sigma, the covariance matrix.}
}
\value{
A \code{matrix} that is a Wishart distribution, aka the sample covariance matrix of a Multivariate Normal Distribution.
}
\description{
Creates a random wishart distribution when given degrees of freedom and a sigma matrix.
}
\seealso{
\code{\link{riwishart}}
}
