% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_likelihood.R
\name{simulate_likelihood}
\alias{simulate_likelihood}
\title{Simulation of the likelihood}
\usage{
simulate_likelihood(logr, n, nsim, sigma_true)
}
\arguments{
\item{logr}{log of r param (dbl)}

\item{n}{Number of steps (int)}

\item{nsim}{Simulation's number (int)}

\item{sigma_true}{The true sigma value (dbl)}
}
\value{
The likelihood (dbl)
}
\description{
Simulation of the likelihood
}
